shimmer-wasp: Only add access nodes to chain registry if chain exists and peers are defined in .env

+++ common/scripts/prepare_docker_functions.sh
@@ -333,19 +333,16 @@ generate_peering_json() {
 
 configure_trusted_peers() {
   local trustedPeersPath="$1"
-  local chainRegistryPath="$2"
-
-  local chainID="${WASP_CHAIN_ADDRESS:-smr1prxvwqvwf7nru5q5xvh5thwg54zsm2y4wfnk6yk56hj3exxkg92mx20wl3s}"
 
   if [ ! -f "${trustedPeersPath}" ]; then echo "{\"trustedPeers\": []}" > "${trustedPeersPath}"; fi
-  if [ ! -f "${chainRegistryPath}" ]; then echo "{\"chainRecords\": [{\"chainID\": \"${chainID}\", \"active\": false}]}" > "${chainRegistryPath}"; fi
 
   if ! grep -q -E "^WASP_TRUSTED_NODE_[0-9]+_URL" .env; then
-    echo "  Skipped: No trusted peers defined in .env"
+    echo "  Cleared all trusted peers: No peers defined in .env"
+    jq '.trustedPeers |= map(select(.name == "me"))' \
+      "${trustedPeersPath}" > "${trustedPeersPath}.tmp" && mv -f "${trustedPeersPath}.tmp" "${trustedPeersPath}"
   else
-    jq '.trustedPeers |= map(select(.name == "me"))' "${trustedPeersPath}" > "${trustedPeersPath}.tmp" && mv -f "${trustedPeersPath}.tmp" "${trustedPeersPath}"
-    set_config "${chainRegistryPath}" ".chainRecords[0].chainID" "\"${chainID}\""
-    set_config "${chainRegistryPath}" ".chainRecords[0].accessNodes" "[]" "suppress"
+    jq '.trustedPeers |= map(select(.name == "me"))' \
+      "${trustedPeersPath}" > "${trustedPeersPath}.tmp" && mv -f "${trustedPeersPath}.tmp" "${trustedPeersPath}"
 
     grep -E "^WASP_TRUSTED_NODE_[0-9]+_URL" .env | sort | while IFS= read -r trustedNodeUrl; do
       trustedNodeNumber=$(echo "${trustedNodeUrl}" | cut -d '_' -f 4)
@@ -358,11 +355,41 @@ configure_trusted_peers() {
       jq -r --arg name "${name}" --arg url "${url}" --arg pubKey "${pubKey}" \
         '.trustedPeers += [{name: $name, publicKey: $pubKey, peeringURL: $url}]' \
         "${trustedPeersPath}" > "${trustedPeersPath}.tmp" && mv -f "${trustedPeersPath}.tmp" "${trustedPeersPath}"
-
-      jq -r --arg pubKey "${pubKey}" \
-        '.chainRecords[0].accessNodes += [$pubKey]' \
-        "${chainRegistryPath}" > "${chainRegistryPath}.tmp" && mv -f "${chainRegistryPath}.tmp" "${chainRegistryPath}"
     done
   fi
-  chown 65532:65532 "${trustedPeersPath}" "${chainRegistryPath}"
+  chown 65532:65532 "${trustedPeersPath}"
+}
+
+configure_chain_access_nodes() {
+  local chainRegistryPath="$1"
+  local evmChainName="$2"
+  local evmChainID="$3"
+
+  if [ ! -f "${chainRegistryPath}" ]; then
+    echo "  Skipped (chain registry): No chain registry found"
+  elif ! jq -e '.chainRecords' "${chainRegistryPath}" >/dev/null || \
+       ! jq -e --arg chainID "${evmChainID}" '.chainRecords[] | select(.chainID == $chainID)' "${chainRegistryPath}" >/dev/null; then
+    echo "  Skipped (chain registry): ${evmChainName} not present in chain registry"
+  elif ! grep -q -E "^WASP_TRUSTED_NODE_[0-9]+_URL" .env; then
+    echo "  Cleared all access nodes: No peers defined in .env"
+    jq --arg chainID "${evmChainID}" '.chainRecords |= map(if .chainID == $chainID then .accessNodes = [] else . end)' \
+      "${chainRegistryPath}" > "${chainRegistryPath}.tmp" && mv -f "${chainRegistryPath}.tmp" "${chainRegistryPath}"
+  else
+    jq --arg chainID "${evmChainID}" '.chainRecords |= map(if .chainID == $chainID then .accessNodes = [] else . end)' \
+      "${chainRegistryPath}" > "${chainRegistryPath}.tmp" && mv -f "${chainRegistryPath}.tmp" "${chainRegistryPath}"
+
+    grep -E "^WASP_TRUSTED_NODE_[0-9]+_URL" .env | sort | while IFS= read -r trustedNodeUrl; do
+      trustedNodeNumber=$(echo "${trustedNodeUrl}" | cut -d '_' -f 4)
+      name=$(grep -E "WASP_TRUSTED_NODE_${trustedNodeNumber}_NAME" .env | cut -d '=' -f 2)
+      if [ -z "${name}" ]; then name="peer${trustedNodeNumber}"; fi
+      pubKey=$(grep -E "WASP_TRUSTED_NODE_${trustedNodeNumber}_PUBKEY" .env | cut -d '=' -f 2)
+
+      echo "  Adding access node ${name}: ${pubKey}"
+      jq --arg chainID "${evmChainID}" --arg access_node "${pubKey}" \
+        '.chainRecords |= map(if .chainID == $chainID then .accessNodes += [$access_node] else . end)' \
+        "${chainRegistryPath}" > "${chainRegistryPath}.tmp" && mv -f "${chainRegistryPath}.tmp" "${chainRegistryPath}"
+    done
+
+    chown 65532:65532 "${chainRegistryPath}"
+  fi
 }
